Rename FOI files (#63)
