+++ pyEpiabm/pyEpiabm/property/personal_foi.py
