+++ pyEpiabm/pyEpiabm/property/spatial_foi.py
