+++ pyEpiabm/pyEpiabm/tests/test_property/test_space_foi.py
